refactoring + manifest processing
