+++ on2logic/plot.py
@@ -1,42 +1,9 @@
 import pandas as  pd
 import numpy as np
 import torch
-import timm
 from scipy import spatial
 import seaborn as sns
-from torchvision import datasets, transforms
 import matplotlib.pyplot as plt
-
-def load_image_model(model_name:str ='restnet50') -> torch.nn.Module:
-    image_model = timm.create_model(model_name, pretrained=True)
-    image_model.eval() # this turns off dropout etc, important!
-    return image_model
-    
-def load_image_folder_and_return_dataset(folder_name:str = '../data/images/', image_size:int = 500):
-
-    transform = transforms.Compose([transforms.Resize(image_size),
-                                    transforms.CenterCrop(image_size-1),
-                                    transforms.ToTensor()])
-    
-    image_dataset = datasets.ImageFolder(folder_name, transform=transform)
-    image_label_to_name =  {value:key for key, value in image_dataset.class_to_idx.items()}
-    
-    page_numbers = [name[0].split('.')[-2].split('-')[-1] for name in image_dataset.imgs] # page numbers numerically not strictly correct
-    return image_dataset, image_label_to_name, page_numbers
-
-def generate_vectors_from_preprocessed_folder(image_model, image_dataset):
-    image_dataloader = torch.utils.data.DataLoader(image_dataset, batch_size=1, shuffle=False)
-    image_vectors = []
-    image_labels = []
-    for batch_of_images, batch_of_labels in image_dataloader:
-        batch_of_vectors = image_model(batch_of_images)
-        image_vectors.append(batch_of_vectors)
-        image_labels.append(batch_of_labels)
-        
-    image_vectors = torch.cat(image_vectors).detach().numpy()
-    image_labels = torch.cat(image_labels).detach().numpy()
-    return image_vectors, image_labels
-
 
 def plot_image_from_index(image_dataset, image_index, gray=True, title=None, ax=None):
     sns.set_style('white')
@@ -60,14 +27,6 @@ def plot_image_from_index(image_dataset, image_index, gray=True, title=None, ax=
         if title is not None:
             ax.set_title(title)
     return
-
-
-def generate_manuscript_dataframe(image_vectors, image_labels, label_dictionary, page_numbers):
-    manuscript_dataframe = pd.DataFrame({'vector': [x for x in image_vectors], 'label': image_labels})
-    manuscript_dataframe['manuscript'] = [label_dictionary[label] for label in manuscript_dataframe['label']]
-    manuscript_dataframe['page'] = page_numbers
-    return manuscript_dataframe
-
 
 def cosine_similarity(vector1, vector2):
     return 1 - spatial.distance.cosine(vector1, vector2)
@@ -183,15 +142,3 @@ def similarity_histogram(image_dataset_to_search,
     plt.show()
     return
     
-    
-def case_study_setup(parent_dirname):
-    image_dirname = parent_dirname+'/data/images'
-    numpy_dirname = parent_dirname+'/notebooks/numpy/'
-    manuscript_dataset, manuscript_label_to_name, manuscript_page_numbers = load_image_folder_and_return_dataset(image_dirname)
-
-    manuscript_vectors = np.load(numpy_dirname+'vectors.npy')
-    manuscript_labels = np.load(numpy_dirname+'labels.npy')
-    
-    manuscript_dataframe = generate_manuscript_dataframe(manuscript_vectors, manuscript_labels, manuscript_label_to_name, manuscript_page_numbers)
-    
-    return manuscript_dataset, manuscript_dataframe
