--- tests/__init__.py
@@ -1,3 +1,0 @@
-from os.path import dirname
-
-ROOT = dirname(dirname(__file__))
